CovarianceMatrix.jl

This module implements covariance matrix estimations for EEG data and encoding of BCI trials in the form of covariance matrices.

Covariance matrices are essential objects in EEG data analysis and BCI data classification.

Resources for Covariance Matrices

For manipulating covariance matrices using Riemannian geometry — see package PosDefManifold.jl.

For machine learning based on Riemannian geometry — see package PosDefManifoldML.jl.

For methods based on the diagonalization of one or more covariance matrices, see package Diagonalizations.jl.

Fourier cross-spectra and coherence matrices are special forms of covariance matrices. For producing them, see FourierAnalysis.jl.

Methods

Function	Description

Eegle.CovarianceMatrix.covmat	Many covariance matrix estimators (2 methods)

Eegle.CovarianceMatrix.encode	Encode all trials in a given EEG recording

📖

Eegle.CovarianceMatrix.covmat — Function

(1) function covmat(X::AbstractMatrix{T};
        covtype = SCM,
        prototype::Union{AbstractMatrix, Nothing} = nothing,
        standardize::Bool = false,
        useBLAS::Bool = true,
        threaded::Bool = true,
        reg::Symbol = :rmt,
        tol::Real = real(T)(1e-6),
        maxiter::Int = 200,
        verbose::Bool = false) 
    where T<:Union{Real, Complex}

(2)  function covmat(𝐗::AbstractVector{<:AbstractArray{T}};
        < same arguments as method (1) > ...

Covariance matrix estimation(s) of a single data matrix (e.g., a trial) X (1) or of a vector of $K$ data matrices 𝐗 (2).

Arguments

(1) X: $N×T$ real data matrix, where $N$ and $T$ denotes the number of samples and channels, respectively

(2) 𝐗: a vector holding $k$ such matrices.

Optional Keyword Arguments

covtype: covariance estimator (default = SCM):

SCM : sample covariance matrix (maximum likelihood)

LShrLW : linear shrinkage estimator of (Ledoit and Wolf, 2004)

NShrLW : non-linear shrinkage estimator of (Ledoit and Wolf, 2020)

:Tyler: Tyler's M-estimator (Tyler, 1987)

:nrTyler: normalized regularized Tyler's M-Estimator (Zhang and Wiesel, 2016)

or any estimator from the CovarianceEstimation.jl package.

prototype: optional matrix to be stacked to the data matrix (or matrices) to form a super-trial — see Appendix I in (Congedo et al., 2017). Default = nothing

standardize: if true, standardize the data matrix(ces) (global mean = 0 and global sd = 1 for each matrix) before estimating the covariance. Default = false

useBLAS: optimize the SCM covariance computations using BLAS. Default = true

threaded: enable multi-threading across the $k$ matrices. For (2) only. Default = true

Only for M-estimators:

tol: the tolerance for the stopping criterion of the iterative algorithm (default = 1e-6)

maxiter: the maximum number of iterations allowed (default = 200)

verbose: if true, information about convergence is printed in the REPL (default = false).

Only for the normalized regularized Tyler's M-Estimator. Default = :rmt:

reg: if it is :rmt, the random matrix theory shrinkage is used, otherwise the Ledoit and Wolf linear shrinkage is used.

Return

(1): The covariance matrix estimation as a Julia Hermitian matrix

(2): A vector of $K$ covariance matrix estimations as an HermitianVector type.

Examples

using Eegle # or using Eegle.CovarianceMatrix

# Method (1)

C = covmat(randn(128, 19)) # sample covariance matrix

C = covmat(randn(128, 19); covtype=LShrLW)

C = covmat(randn(128, 19); covtype=:Tyler)

# Method (2)

𝐗 = [randn(128, 19) for k=1:10]

𝐂 = covmat(𝐗)

C = covmat(𝐗; covtype=LShrLW)

C = covmat(𝐗; covtype=:Tyler)

Eegle.CovarianceMatrix.encode — Function

    function encode(o::EEG, paradigm::Symbol;
        covtype=SCM,
        targetLabel::String = "",
        overlapping::Bool = false,
        weights = :a,
        pcadim::Int = 8,
        standardize::Bool = false,
        tikh = 0,
        useBLAS = true,
        threaded = true,
        reg::Symbol = :rmt,
        tol::Real = real(T)(1e-6),
        maxiter::Int = 200,
        verbose::Bool = false)

Encode all trials in an EEG recording as covariance matrices for a given BCI paradigm. This is used in Riemannian geometry machine learning. The supported BCI paradigms are Motor Imagery (MI), Event-Related Potentials (ERP) and P300. For details, see Appendix I in (Congedo et al., 2017).

Arguments

o: an instance of the EEG data structure containing trials and metadata

paradigm: BCI paradigm, either :ERP, :P300, or :MI:

for :ERP, prototypes for all classes are stacked and covariance is computed on super-trials

for :P300, only the target class prototype is stacked

for :MI, no prototype is used; covariance is computed on the trial as it is.

Optional Keyword Arguments

covtype, useBLAS, tol, maxiter and verbose — see Eegle.CovarianceMatrix.covmat, to which they are passed.

targetLabel: mandatory label of the target class (P300 paradigm only, usually: "target")

overlapping: for prototype mean ERP estimations (ERP/P300 only). Default = false:

if true, use multivariate regression

if false, use the arithmetic average — see mean.

weights: weights for prototype mean ERP estimations (ERP/P300 only). Default = :a — see mean

pcadim: number of PCA components of the prototype. They replace the prototype (ERP/P300 only, default = 0, which does not apply PCA)

standardize: standardize trials and prototype (global mean 0 and sd 1) before covariance estimation (default: false)

tikh: Tikhonov regularization parameter (0, the default, does not apply regularization). It is applied after covariance estimation

threaded: enable multi-threaded covariance estimations across trials (default: true).

Throw

ArgumentError if class label targetLabel is not found in o.clabels (for P300 paradigm).

ArgumentError if paradigm is not one of :ERP, :P300, or :MI.

Return A vector of $k$ covariance matrix estimations as a HermitianVector type.

Examples

using Eegle # or using Eegle.CovarianceMatrix
xxx