Miscellaneous.jl

This module implements miscellaneous utilities.

Function	Description

Eegle.Miscellaneous.minima	local minuma of a sequence

Eegle.Miscellaneous.maxima	local maxima of a sequence

Eegle.Miscellaneous.waste	garbage collect any number of arguments

📖

Eegle.Miscellaneous.minima — Function

function minima(v::AbstractVector{T}) 
where T<:Real

Return the 2-tuple formed by the vector of local minima of vector v and the vector of the indices of v corresponding to the minima.

This is useful in several situations. For example, Eegle uses it to segment spontaneous EEG data (see Eegle.Processing.epoching).

Eegle.Miscellaneous.maxima — Function

function maxima(v::AbstractVector{T}) 
where T<:Real

Return the 2-tuple formed by the vector of local maxima of vector v and the vector of the indices of v corresponding to the maxima.

Eegle.Miscellaneous.waste — Function

function waste(args...)

Force garbage collector to free memory for all arguments passed as args....

The arguments can be of any type. See here

Examples

using Eegle # or using Eegle.Miscellaneous

A = randn(1000, 1000)
b = randn(10000)

waste(A, b)