Preprocessing.jl

This module implements preprocessing for EEG data.

See also Processing.jl

Methods

Function	Description

Eegle.Preprocessing.resample	resample EEG data

Eegle.Preprocessing.standardize	standardize EEG data

Eegle.Preprocessing.removeChannels	remove channels from EEG data

Eegle.Preprocessing.removeSamples	remove samples from EEG data

Eegle.Preprocessing.embedLags	lag-embedding of EEG data

📖

StatsBase.standardize — Function

    function standardize(X::AbstractArray{T}; 
        robust = false,
        prop::Real = 0.2) 
    where T<:Real

Standardize the whole $T×N$ EEG recording X, where $T$ and $N$ denotes the number of samples and channels (sensors), respectively, using:

the arithmetic mean and standard deviation if robust is false (default)

the Winsorized (trimmed) mean and standard deviation if robust is true.

The trimmed statistics are computed excluding the prop proportion of data at both sides (default=0.2), thus, prop is used only if robust is true.

Example

using Eegle # or using Eegle.Preprocessing

X=randn(1024, 19)

stX=standardize(X)

stX=standardize(X; robust=true, prop=0.1)

DSP.Filters.resample — Function

    function resample(  X::AbstractMatrix{T},
                        sr::S,
                        rate::Union{T, S, Rational};
        rel_bw::Float64 = 1.0,
        attenuation::Int = 60,
        stim::Union{Vector{S}, Nothing} = nothing) 
    where {T<:Real, S<:Int}

Resampling of the $T×N$ EEG recording X, where $T$ and $N$ denotes the number of samples and channels (sensors), respectively, using the polyphase FIR filter with Kaiser window filter taps, as per the resample method in DSP.jl.

For kwarg rel_bw and attenuation — see resample.

If a stimulation vector stim is passed as kwarg, It will be resampled so as to match the resampling of X as precisely as possible. stim must be a vector of $T$ integers.

Resampling

If you need to work with individual trials, do not resample trials individually; rather, resample the whole EEG recording and then extract the trials. Function Eegle.InOut.readNY allows you to do resampling and extract trials this way.

Real resampling rate

The use of real resampling rates need to be tested

Examples

using Eegle # or using Eegle.Preprocessing

sr = 128
X = randn(sr*10, 19)

Y=resample(X, sr, 1//4) # downsample by a factor 4

Y=resample(X, sr, 2) # upsample by a factor 2, i.e., double the sampling rate

Eegle.Preprocessing.removeChannels — Function

    function removeChannels(X::AbstractMatrix{T}, 
                            what::Union{Int, Vector{S}},
                            sensors::Vector{String}) 
    where {T<:Real, S<:Int}

Remove one or more channels, i.e., columns, from the $T×N$ EEG recording X, where $T$ and $N$ denotes the number of samples and channels (sensors), respectively, and remove the corresponding elements from sensors, the provided associated vector of $N$ sensor labels.

For the use of kwarg what — see method Eegle.Miscellaneous.remove, which can be used if all you need is to remove channels from X.

Return the 3-tuple (newX, s, ne), where newX is the new EEG recording, s is the new sensor labels vector and ne is the new number of channels (sensors) in newX.

See Also Eegle.InOut.readSensors

Examples

using Eegle # or using Eegle.Preprocessing

# xxx load data

# remove second channel
X, sensors, ne = removeChannels(X, 2, sensors)

# remove the first five channels
X, sensors, ne = removeChannels(X, collect(1:5), sensors)

# remove the channel labeled as 'Cz' in `sensors`
X, sensors, ne = removeChannels(X, findfirst(x->x=="Cz", sensors), sensors)

Eegle.Preprocessing.removeSamples — Function

    function removeSamples( X::AbstractMatrix{T}, 
                            what::Union{Int, Vector{S}},
                            stim::Vector{String}) 
    where {T<:Real, S<:Int}

Remove one or more samples, i.e., rows, from the $T×N$ EEG recording X, where $T$ and $N$ denotes the number of samples and channels (sensors), respectively, and remove the corresponding elements from stim, the associated stimulation vector.

For the use of kwarg what — see method Eegle.Miscellaneous.remove, which can be used instead if all you need is to remove samples from X.

Print a warning if elements in what correspond to non-zero tags in stim.

Return the 3-tuple (newX, s, ne), where newX is the new data, s is the new stimulation vector and ns is the new number of samples in newX.

Examples

using Eegle

# load data xxx

# remove second sample
X, stim, ne = removeSamples(X, 2, stim)

# remove the first 128 samples
X, stim, ne = removeSamples(X, collect(1:128), stim)

# remove every other sample
X, stim, ne = removeSamples(X, collect(1:2:length(stim)), stim)

Eegle.Preprocessing.embedLags — Function

    function embedLags( X::AbstractMatrix{T}, 
                        lags = 0) 
    where T<:Real 

Lag-embedding is a technique to augment the data. Second-order statistics of the augmented data, that is, covariance or cross-spectral matrices, hold information not only of volume condution and instantaneous connectivity, but also of lagged connectivity. These matrices can be used, for example, in blind source separation and Riemannian classification.

Given the $T×N$ EEG recording X, where $T$ and $N$ denotes the number of samples and channels (sensors), respectively, and $L>0$ lags, return the $T×N(L+1)$ lag-embedded data matrix

\[X_{\text{lags}} = \left[ X^{(0)} \;\; X^{(1)} \;\; \cdots \;\; X^{(L)} \right]\]

where, letting $\mathbf{0}_{A \times B}$ the $A \times B$ matrix of zeros, for each $l = 0, \ldots, L$, the $T×N$ lagged partition $X^{(l)}$ is defined as

\[X^{(l)} = \left[ \begin{matrix} \mathbf{0}_{(L - l) \times N} \\ X[1:(T - L), \: :] \\ \mathbf{0}_{l \times N} \end{matrix} \right].\]

Notice that the are no zeros appended to the first partition and no zeros prepended to the last partition. Notice also that the lag-embedded data has the same size of the input, however the last $L$ samples are lost.

Example

using Eegle # or using Eegle.Preprocessing

X=randn(1024, 19)

elX=emdedLags(X, 4)